fix(codegen-csharp): a payload's field.decimal is `decimal`, not `double` (#309)
`PayloadCodegen`'s scalar map sent `field.decimal` to `double`, so the payload
tier silently rounded — on the shape most likely to carry money.

Every other tier and port already disagreed with it. This port's own entity
generator maps it to `decimal` (`CSharpNaming`), ADR-0019 pins the runtime return
type to `decimal`, Kotlin and Java use `BigDecimal` (their mapper's comment reads
"NEVER a float"), and TypeScript uses `string` because its Postgres driver returns
NUMERIC as a string specifically to avoid float rounding. So this was not an open
question about precision — it was one map disagreeing with a settled contract.

The test asserting `double` is corrected rather than regenerated: it pinned the
defect the same way the `required` assertions did, which is why both shipped.

Adopter-visible: a payload property typed `double` becomes `decimal`.

C# 1592 green; gates 10/10.

+++ server/csharp/MetaObjects.Codegen/PayloadCodegen.cs
@@ -58,7 +58,14 @@ public static class PayloadCodegen
             [FIELD_SUBTYPE_CURRENCY]  = "long",
             [FIELD_SUBTYPE_DOUBLE]    = "double",
             [FIELD_SUBTYPE_FLOAT]     = "double",
-            [FIELD_SUBTYPE_DECIMAL]   = "double",
+            // #309 (second observation) — `decimal`, NOT `double`. field.decimal is
+            // precision-exact, and every other tier and port already says so: this port's
+            // entity generator maps it to `decimal` (CSharpNaming), ADR-0019 pins the
+            // runtime return type to `decimal`, Kotlin/Java use BigDecimal ("NEVER a
+            // float"), and TypeScript uses `string` because its pg driver returns NUMERIC
+            // as a string to avoid float rounding. Only the payload tier said `double`,
+            // which silently rounds — the one shape most likely to carry money.
+            [FIELD_SUBTYPE_DECIMAL]   = "decimal",
             [FIELD_SUBTYPE_BOOLEAN]   = "bool",
             [FIELD_SUBTYPE_DATE]      = "string",
             [FIELD_SUBTYPE_TIME]      = "string",
